+++ ocaml-eglot.el
@@ -353,12 +353,16 @@ If there is no available holes, it returns the first one of HOLES."
 
 (defun ocaml-eglot--phrase (direction)
   "Move to the next or previous phrase using DIRECTION."
-  (let* ((result (ocaml-eglot-req--phrase direction))
-         (json-result (ocaml-eglot-util--merlin-call-result result))
-         (pos (cl-getf json-result :pos)))
-    (when pos
-      (let ((target (ocaml-eglot-util--pos-to-point pos)))
-        (ocaml-eglot-util--goto-char target)))))
+  (if (ocaml-eglot-req--server-capable :experimental :ocamllsp :handlePhrase)
+      (let ((result (ocaml-eglot-req--phrase direction)))
+        (when result
+          (ocaml-eglot-util--jump-to result)))
+    (let* ((result (ocaml-eglot-req--phrase-legacy direction))
+           (json-result (ocaml-eglot-util--merlin-call-result result))
+           (pos (cl-getf json-result :pos)))
+      (when pos
+        (let ((target (ocaml-eglot-util--pos-to-point pos)))
+          (ocaml-eglot-util--goto-char target))))))
 
 (defun ocaml-eglot-phrase-next ()
   "Go to the beginning of the next phrase."
@@ -540,9 +544,13 @@ It use the ARG to use local values or not."
 (defun ocaml-eglot-type-expression (expression)
   "Prompt the user for expression EXPRESSION and print its type."
   (interactive "sExpression: ")
-  (let* ((result (ocaml-eglot-req--type-expression expression))
-         (type-expr (ocaml-eglot-util--merlin-call-result result)))
-    (ocaml-eglot-type-enclosing--display type-expr nil)))
+  (if (ocaml-eglot-req--server-capable
+       :experimental :ocamllsp :handleTypeExpression)
+      (let ((type-expr (ocaml-eglot-req--type-expression expression)))
+        (ocaml-eglot-type-enclosing--display type-expr nil))
+    (let* ((result (ocaml-eglot-req--type-expression-legacy expression))
+           (type-expr (ocaml-eglot-util--merlin-call-result result)))
+      (ocaml-eglot-type-enclosing--display type-expr nil))))
 
 (defun ocaml-eglot-type-enclosing (&optional prefix)
   "Print the type of the expression under point (or of the region, if it exists).